[restructure] move email related utils to lib/utils/mail
